Checkout screen: caricati indirizzo e metodo correnti

+++ app/src/main/java/com/example/grocify/api/GeocodingResponse.kt
@@ -73,5 +73,5 @@ data class TopLeftPoint(
 )
 
 data class MatchConfidence(
-    val score: Int
+    val score: Double
 )
